+++ pkg/framework/objects/athena_credential/datasource.go
@@ -1,0 +1,106 @@
+package athena_credential
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ datasource.DataSource              = &athenaCredentialDataSource{}
+	_ datasource.DataSourceWithConfigure = &athenaCredentialDataSource{}
+)
+
+// NewAthenaCredentialDataSource is a helper function to simplify the provider implementation.
+func NewAthenaCredentialDataSource() datasource.DataSource {
+	return &athenaCredentialDataSource{}
+}
+
+// athenaCredentialDataSource is the data source implementation.
+type athenaCredentialDataSource struct {
+	client *dbt_cloud.Client
+}
+
+// Configure adds the provider configured client to the data source.
+func (d *athenaCredentialDataSource) Configure(
+	ctx context.Context,
+	req datasource.ConfigureRequest,
+	resp *datasource.ConfigureResponse,
+) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	client, ok := req.ProviderData.(*dbt_cloud.Client)
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Data Source Configure Type",
+			fmt.Sprintf(
+				"Expected *dbt_cloud.Client, got: %T. Please report this issue to the provider developers.",
+				req.ProviderData,
+			),
+		)
+		return
+	}
+
+	d.client = client
+}
+
+// Metadata returns the data source type name.
+func (d *athenaCredentialDataSource) Metadata(
+	ctx context.Context,
+	req datasource.MetadataRequest,
+	resp *datasource.MetadataResponse,
+) {
+	resp.TypeName = req.ProviderTypeName + "_athena_credential"
+}
+
+// Schema defines the schema for the data source.
+func (d *athenaCredentialDataSource) Schema(
+	ctx context.Context,
+	req datasource.SchemaRequest,
+	resp *datasource.SchemaResponse,
+) {
+	resp.Schema = datasourceSchema
+}
+
+// Read refreshes the Terraform state with the latest data.
+func (d *athenaCredentialDataSource) Read(
+	ctx context.Context,
+	req datasource.ReadRequest,
+	resp *datasource.ReadResponse,
+) {
+	var state AthenaCredentialDataSourceModel
+	diags := req.Config.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	projectID := int(state.ProjectID.ValueInt64())
+	credentialID := int(state.CredentialID.ValueInt64())
+
+	credential, err := d.client.GetAthenaCredential(projectID, credentialID)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error reading Athena credential",
+			"Could not read Athena credential ID "+state.ID.ValueString()+": "+err.Error(),
+		)
+		return
+	}
+
+	// Map response body to model
+	state.ID = types.StringValue(fmt.Sprintf("%d:%d", projectID, credentialID))
+	state.Schema = types.StringValue(credential.UnencryptedCredentialDetails.Schema)
+
+	// Set state
+	diags = resp.State.Set(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+}